Used common components for both `suggestion` and `bug` feedback pages
