+++ frontend/src/pages/Feedback/Suggestion.page.tsx
@@ -1,0 +1,5 @@
+import FeedbackSubmitPage from '#modules/feedback/view/shared/FeedbackSubmitPage';
+
+export default function SuggestionPage() {
+  return <FeedbackSubmitPage type="suggestion" />;
+}
